+++ changelog.md
@@ -3,6 +3,7 @@
 ## 1.12.1
 * Spawning feature (renamed from Base) can now be disabled
 * Creeper cena now also breaks more blocks if SR is present
+* Rebalanced some default values
 
 ## 1.12.0
 * Added Biting Zombies feature