+++ changelog.md
@@ -5,6 +5,7 @@
 * Launch creepers have now reduced particles
 * Fisher and Pearler zombies no longer use the item if in water
 * Fisher zombies cooldown reduced
+* Mobs no longer flee from TNTs
 
 ## 1.12.1
 * Spawning feature (renamed from Base) can now be disabled